drainase 2022

+++ app/Http/Controllers/Drainase2022Controller.php
@@ -32,7 +32,7 @@ class Drainase2022Controller extends Controller
                 })
                 ->addColumn('action', function ($row) {
 
-                    $btn = '  <a href="javascript:void(0);" class="btn btn-sm btn-primary"><i class="la la-pencil"></i></a>
+                    $btn = '  <a href="' . route('drainase2022.edit', $row->id) . '" class="btn btn-sm btn-primary"><i class="la la-pencil"></i></a>
                                                     <a href="javascript:void(0);" onclick="destroy(' . $row->id . ')" class="btn btn-sm btn-danger"><i class="la la-trash-o"></i></a>';
                     return $btn;
                 })
@@ -53,6 +53,16 @@ class Drainase2022Controller extends Controller
         return view('pages.drainase2022.form', compact(['title', 'kecamatan', 'kelurahan', 'jalan', 'fisik']));
     }
 
+    public function edit($id)
+    {
+        $drainase2022 = Drainase2022::find($id);
+        $title = 'Edit Data';
+        $kecamatan = Kecamatan::all();
+        $kelurahan = Kelurahan::all();
+        $jalan = Jalan::all();
+        $fisik = Kategori::where('induk', 'Kondisi Fisik')->get();
+        return view('pages.drainase2022.form', compact(['title', 'kecamatan', 'kelurahan', 'jalan', 'fisik', 'drainase2022']));
+    }
 
     public function store(Request $request)
     {
@@ -70,7 +80,7 @@ class Drainase2022Controller extends Controller
             'tipe' =>  'required',
             'kondisi_fisik' =>  'required',
             'foto' => 'image|mimes:jpeg,jpg,png|max:5120|nullable',
-            'file_kmz' =>  ['required', 'max:5000', function ($attribute, $value, $fail) use ($request) {
+            'file_kmz' =>  [$request->file_kmz_edit != '' ? '' : 'required', 'max:5000', function ($attribute, $value, $fail) use ($request) {
                 if ($request->file('file_kmz')->getClientOriginalExtension() != 'kmz') {
                     $fail('The file kmz must be a file of type: kmz.');
                 }
@@ -81,29 +91,35 @@ class Drainase2022Controller extends Controller
             return response()->json(['error' => $validator->errors(), 'status' => false]);
         } else {
             if (empty($request->id)) { //CREATE
-                $drainase2020 = new Drainase2022();
-                $drainase2020->kode_saluran = $request->kode_saluran;
-                $drainase2020->kecamatan = $request->kecamatan;
-                $drainase2020->kelurahan = $request->kelurahan;
-                $drainase2020->nama_jalan = $request->nama_jalan;
-                $drainase2020->sisi = $request->sisi;
-                $drainase2020->panjang = $request->panjang;
-                $drainase2020->tinggi = $request->tinggi;
-                $drainase2020->lebar_atas = $request->lebar_atas;
-                $drainase2020->lebar_bawah = $request->lebar_bawah;
-                $drainase2020->arah = $request->arah;
-                $drainase2020->tipe = $request->tipe;
-                $drainase2020->kondisi_fisik = $request->kondisi_fisik;
-                $drainase2020->foto = $request->foto;
+                $drainase2022 = new Drainase2022();
+                $drainase2022->kode_saluran = $request->kode_saluran;
+                $drainase2022->kecamatan = $request->kecamatan;
+                $drainase2022->kelurahan = $request->kelurahan;
+                $drainase2022->nama_jalan = $request->nama_jalan;
+                $drainase2022->sisi = $request->sisi;
+                $drainase2022->panjang = $request->panjang;
+                $drainase2022->tinggi = $request->tinggi;
+                $drainase2022->lebar_atas = $request->lebar_atas;
+                $drainase2022->lebar_bawah = $request->lebar_bawah;
+                $drainase2022->arah = $request->arah;
+                $drainase2022->tipe = $request->tipe;
+                $drainase2022->kondisi_fisik = $request->kondisi_fisik;
 
                 DB::beginTransaction();
                 try {
                     if ($request->hasFile('file_kmz')) {
-                        $name = strtolower(str_replace(' ', '', $request->kelurahan)) . '/' . $request->kode_saluran . '.' . $request->file('file_kmz')->getClientOriginalExtension();
-                        $drainase2020->file_kmz = $name;
-                        $request->file('file_kmz')->storeAs('kmz', $name, 'public');
+                        $kmzName = strtolower(str_replace(' ', '', $request->kelurahan)) . '/' . $request->kode_saluran . '.' . $request->file('file_kmz')->getClientOriginalExtension();
+                        $drainase2022->file_kmz = $kmzName;
+                        $request->file('file_kmz')->storeAs('kmz', $kmzName, 'public');
                     }
-                    $drainase2020->save();
+
+                    if ($request->hasFile('foto')) {
+                        $fotoName =  time() . rand() . '.' . $request->file('foto')->getClientOriginalExtension();
+                        $drainase2022->foto = $fotoName;
+                        $request->file('foto')->storeAs('foto', $fotoName, 'public');
+                    }
+
+                    $drainase2022->save();
                     DB::commit();
                     return response()->json(['status' => true, 'redirect' =>  route("drainase2022.create")]);
                 } catch (\Exception $e) {
@@ -111,25 +127,40 @@ class Drainase2022Controller extends Controller
                     // return response()->json(['status' => false, 'err' => $e->getMessage()]);
                 }
             } else { //UPDATE
-                $drainase2020 =  Drainase2022::find($request->id);
-                $drainase2020->kode_saluran = $request->kode_saluran;
-                $drainase2020->kecamatan = $request->kecamatan;
-                $drainase2020->kelurahan = $request->kelurahan;
-                $drainase2020->nama_jalan = $request->nama_jalan;
-                $drainase2020->sisi = $request->sisi;
-                $drainase2020->panjang = $request->panjang;
-                $drainase2020->tinggi = $request->tinggi;
-                $drainase2020->lebar_atas = $request->lebar_atas;
-                $drainase2020->lebar_bawah = $request->lebar_bawah;
-                $drainase2020->arah = $request->arah;
-                $drainase2020->tipe = $request->tipe;
-                $drainase2020->kondisi_fisik = $request->kondisi_fisik;
-                $drainase2020->foto = $request->foto;
-                $drainase2020->file_kmz = $request->file_kmz;
+                $drainase2022 =  Drainase2022::find($request->id);
+                $drainase2022->kode_saluran = $request->kode_saluran;
+                $drainase2022->kecamatan = $request->kecamatan;
+                $drainase2022->kelurahan = $request->kelurahan;
+                $drainase2022->nama_jalan = $request->nama_jalan;
+                $drainase2022->sisi = $request->sisi;
+                $drainase2022->panjang = $request->panjang;
+                $drainase2022->tinggi = $request->tinggi;
+                $drainase2022->lebar_atas = $request->lebar_atas;
+                $drainase2022->lebar_bawah = $request->lebar_bawah;
+                $drainase2022->arah = $request->arah;
+                $drainase2022->tipe = $request->tipe;
+                $drainase2022->kondisi_fisik = $request->kondisi_fisik;
                 try {
-                    $drainase2020->save();
+                    if ($request->hasFile('file_kmz')) {
+                        if ($drainase2022->file_kmz != '' && Storage::exists('public/kmz/' . $request->file_kmz)) {
+                            Storage::delete('public/kmz/' . $request->file_kmz);
+                        }
+                        $kmzName = strtolower(str_replace(' ', '', $request->kelurahan)) . '/' . $request->kode_saluran . '.' . $request->file('file_kmz')->getClientOriginalExtension();
+                        $drainase2022->file_kmz = $kmzName;
+                        $request->file('file_kmz')->storeAs('kmz', $kmzName, 'public');
+                    }
+                    if ($request->hasFile('foto')) {
+                        if ($drainase2022->foto != '' && Storage::exists('public/foto/' . $request->foto)) {
+                            Storage::delete('public/foto/' . $request->foto);
+                        }
+                        $fotoName =  time() . rand() . '.' . $request->file('foto')->getClientOriginalExtension();
+                        $drainase2022->foto = $fotoName;
+                        $request->file('foto')->storeAs('foto', $fotoName, 'public');
+                    }
+
+                    $drainase2022->save();
                     DB::commit();
-                    return response()->json(['status' => true, 'redirect' =>  route("drainase2020.edit", $request->id)]);
+                    return response()->json(['status' => true, 'redirect' =>  route("drainase2022.edit", $request->id)]);
                 } catch (\Throwable $th) {
                     DB::rollBack();
                 }
@@ -161,13 +192,18 @@ class Drainase2022Controller extends Controller
     public function export()
     {
         return Excel::download(new Drainase2022Export, 'Drainase2022.xlsx');
+
+        return redirect()->back();
     }
 
     public function destroy($id)
     {
         $drainase2022 = Drainase2022::findOrFail($id);
-        if (Storage::exists('public/kmz/' . $drainase2022->file_kmz)) {
+        if ($drainase2022->file_kmz != '' && Storage::exists('public/kmz/' . $drainase2022->file_kmz)) {
             Storage::delete('public/kmz/' . $drainase2022->file_kmz);
+        }
+        if ($drainase2022->foto != '' && Storage::exists('public/foto/' . $drainase2022->foto)) {
+            Storage::delete('public/foto/' . $drainase2022->foto);
         }
         $drainase2022->delete();
         return response()->json(['status' => true]);
